fix(create): remove tests, move/rename some files
