--- packages/create-purplet/coolio/.prettierignore
@@ -1,4 +1,0 @@
-dist
-.purplet
-LICENSE
-build
