--- packages/create-purplet/coolio/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "create-purplet",
-  "version": "1.0.0",
-  "private": true,
-  "type": "module",
-  "main": "dist",
-  "scripts": {
-    "build": "purplet build",
-    "build-http": "purplet build-http",
-    "dev": "purplet dev",
-    "format": "prettier --write \"**/*.{ts,ts,js,md,json}\"",
-    "guild-manager": "purplet guild-manager"
-  },
-  "dependencies": {
-    "prettier": "^2.7.1",
-    "purplet": "next"
-  },
-  "optionalDependencies": {
-    "bufferutil": "^4.0.6",
-    "erlpack": "^0.1.4",
-    "utf-8-validate": "^5.0.9",
-    "zlib-sync": "^0.1.7"
-  }
-}
