--- packages/create-purplet/coolio/purplet.config.js
@@ -1,2 +1,0 @@
-const config = {};
-export default config;
