--- packages/create-purplet/coolio/tsconfig.json
@@ -1,3 +1,0 @@
-{
-  "extends": "./.purplet/tsconfig.json"
-}
